Merge upstream

+++ input/drivers_joypad/mfi_joypad.m
@@ -138,7 +138,7 @@ static void apple_gamecontroller_joypad_connect(GCController *controller)
        ? desired_index : 0;
     
     /* prevent same controller getting set twice */
-    if (mfi_controllers[desired_index] != controller.hash)
+    if (mfi_controllers[desired_index] != (uint32_t)controller.hash)
     {
         /* desired slot is unused, take it */
         if (!mfi_controllers[desired_index])
